[Add] game sounds

+++ Header/Sound/SoundService.h
@@ -11,7 +11,7 @@ namespace Sound
     class SoundService
     {
     private:
-        const int background_music_volume = 30;
+        const float background_music_volume = 30.f;
 
         sf::Music background_music;
         sf::Sound sound_effect;
